+++ react/Concept.tsx
@@ -3,11 +3,11 @@ import { Query } from 'react-apollo'
 import { ApolloError } from 'apollo-client'
 import { useRuntime } from 'vtex.render-runtime'
 
-import { IO_DOCUMENTATION } from './utils/constants'
+import { IO_DOCUMENTATION } from './modules/constantExports'
 import DocsRenderer from './components/DocsRenderer'
 import Skeleton from './components/Skeleton'
 import EmptyDocs from './components/EmptyDocs'
-import { maybeAddMdExtension } from './utils'
+import { maybeAddMdExtension } from './modules'
 import MarkdownFile from './graphql/markdownFile.graphql'
 
 const Concept: FC = () => {
