+++ react/ConceptList.tsx
@@ -3,8 +3,8 @@ import { FormattedMessage } from 'react-intl'
 import { useQuery } from 'react-apollo'
 import { Link } from 'vtex.render-runtime'
 
-import { IO_DOCUMENTATION } from './utils/constants'
-import { slug } from './utils'
+import { IO_DOCUMENTATION } from './modules/constantExports'
+import { slug } from './modules'
 import ConceptListQuery from './graphql/conceptList.graphql'
 import Skeleton from './components/Skeleton'
 import UnderConstruction from './UnderConstruction'
