+++ react/components/CustomTags.tsx
@@ -6,48 +6,34 @@ import { useDevice } from 'vtex.device-detector'
 import { Link } from 'vtex.styleguide'
 
 import ArticleNav from './ArticleNav'
-import { slug } from '../utils'
+import { slug } from '../modules'
+import {
+  findDocsIgnoreNodes,
+  removeIgnoredNodesFromDocs,
+} from '../modules/ignoreTokens'
 import CodeBlock from './CodeBlock'
-
-function findContributorsHeadingNode(childNodes: any[]) {
-  return childNodes.findIndex(child => {
-    const childChildrenNodes = child.props.children
-
-    if (!childChildrenNodes || childChildrenNodes.length === 0) {
-      return false
-    }
-
-    const isHeading2Node =
-      child.key.includes('heading') && child.props.level === 2
-
-    return (
-      isHeading2Node && child.props.children[0].props.value === 'Contributors'
-    )
-  })
-}
 
 export const CustomRenderers = {
   root: ({ children }: { children: any[] }) => {
     /* eslint-disable react-hooks/rules-of-hooks */
     const { isMobile } = useDevice()
-    const allContributorsHeadingNodeIndex = findContributorsHeadingNode(
+    const { start, end } = findDocsIgnoreNodes(children)
+
+    const childrenWithoutIgnoredNodes = removeIgnoredNodesFromDocs(
+      start,
+      end,
       children
     )
-
-    const childrenWithoutContributors =
-      allContributorsHeadingNodeIndex > -1
-        ? children.slice(0, allContributorsHeadingNodeIndex)
-        : children
 
     if (isMobile) {
       return (
         <div className="flex flex-column center" style={{ maxWidth: '95vw' }}>
-          {childrenWithoutContributors}
+          {childrenWithoutIgnoredNodes}
         </div>
       )
     }
 
-    const TOCLines = childrenWithoutContributors.reduce(
+    const TOCLines = childrenWithoutIgnoredNodes.reduce(
       (acc: string[], { key, props }: any) => {
         // Skip non-headings and only use H2s
         if (key.indexOf('heading') !== 0 || props.level !== 2) {
@@ -63,7 +49,7 @@ export const CustomRenderers = {
     return (
       <div className="flex justify-between">
         <div className="flex flex-column w-80-l" style={{ maxWidth: '680px' }}>
-          {childrenWithoutContributors}
+          {childrenWithoutIgnoredNodes}
         </div>
         {TOCLines.length > 0 && (
           <div
