+++ react/modules/constantExports.ts
@@ -1,3 +1,5 @@
 export const IO_DOCUMENTATION = 'vtex.io-documentation@0.x'
 export const CSE_CX = '017314625344810072686:u6y7vlxadyu'
 export const CSE_KEY = 'AIzaSyBgix5AaPDcyu6QbboYdAYjQPsOJtZLOdw'
+export const DOCS_IGNORE_START = '<!-- DOCS-IGNORE:start -->'
+export const DOCS_IGNORE_END = '<!-- DOCS-IGNORE:end -->'
